refactor: atualizar models e rotas, remover controllers/rotas em maiúscula
