+++ src/models/User.ts
@@ -7,7 +7,6 @@ enum TypeUser {
 }
 // Define the IUser interface extending mongoose.Document to ensure correct typing for Mongoose documents.
 interface IUser extends mongoose.Document {
-  id: mongoose.Types.ObjectId;
   avatar: string;
   name: string;
   email: string;
@@ -20,7 +19,6 @@ interface IUser extends mongoose.Document {
 // Define the schema for the User model
 const userSchema = new mongoose.Schema(
   {
-    id: { type: mongoose.Schema.Types.ObjectId },
     avatar: { type: String },
     name: { type: String, required: [true, 'name is required'] },
     email: { type: String, required: [true, 'email is required'], unique: true },
