+++ src/routes/establishmentRouter.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import EstablishmentController from '../controllers/establishmentController';
 import { establishmentsUpload } from '../config/multer';
 import { addressRouter } from './addressRouter';
-import { orderRouter } from './OrderRouter';
+import { orderRouter } from './orderRouter';
 
 // Creating a new router instance
 export const establishmentRouter = Router();
