+++ src/routes/index.ts
@@ -3,7 +3,7 @@ import * as path from 'path';
 import { establishmentRouter } from './establishmentRouter';
 import { productsRouter } from './productRouter';
 import { userRouter } from './userRouter';
-import { orderRouter } from './OrderRouter';
+import { orderRouter } from './orderRouter';
 
 // Function to define API routes
 const routes = (app: Router) => {
