+++ src/routes/orderRouter.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import OrderController from '../controllers/OrderController';
+import OrderController from '../controllers/orderController';
 
 export const orderRouter = Router();
